Split and fix all tests
Split the test module into several distinct modules, use underscore
names for test methods and fix calling GmpCommandFactory.
